Made changes to Biting Cold pack
